fastapi/uvicorn/requirements.txt
middleware-apm>=2.0.0
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv=1.0.1